requirements.txt
click==8.1.7
coloredlogs==15.0.1
pymysql==1.1.1
python_dateutil==2.8.2
Requests==2.32.3
